tmp fix image saving and deadlocks
Scene example can now run for extended periods of time.
Added a timeout, upon which request to save frame is sent again.
Furthermore, in order to fix performance hit due to having several images in a given directory, they are split across directories.
For instance, images 0 to 99 are saved in 000/ and 100 to 199 in 100/.
These directories still have to be created prior to launching the scene example demo.
This may become irrelevant with asynchronous batch image saving.