hotfix sync issues in scene generation
As of right now it is impossible to know exactly when the world is done changing.
Furthermore, VisualPlugins are loaded twice per object, due to a suspected bug in Gazebo itself.
Sync issue is fixed with hardcoded delay of 100 ms.
Have generated 100 images in under 3:30 mins.
Performance should scale linearly.